requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
djangorestframework==3.12.2
pytest==6.2.2
pytest-django==4.1.0
pytest-cov==2.11.1